+++ src/generator/mod.rs
@@ -7,7 +7,6 @@ pub mod nbts;
 pub mod buildings;
 pub mod buildings_v2;
 pub mod ships;
-pub mod ships_v2;
 pub mod city_houses;
 pub mod data;
 pub mod paths;
